reimplement omit --callers/--callees
The omit command hides the node, regardless of how many edges it has.
This is a problem when using the command omit --callees X, because it deletes
X and all its callees. This is not what we want, as the callees might have
other callers pointing to them.

Hide the edges instead, and only hide a node if it ends up with no
callers and no callees.

Fixes: #1